refactor: separate error helpers
closes #964
